// src/app/model/meal.model.ts
export interface Meal {
    id: number;
    name: string;
    type: string;
}
export interface food {
    strMeal: string;
    strMealThumb: string;
    idMeal: any;
    price: number;
}
export interface cart {
    strMeal: string;
    strMealThumb: string;
    idMeal: any;
    price: number;
    quantity: number;
}